backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "trinity-backend"
version = "0.1.0"
description = "Trinity Counselor — multi-agent AI relationship counseling system"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.9",

    # Pydantic
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",

    # Anthropic / LLM
    "anthropic>=0.40.0",
    "langchain-anthropic>=0.3.0",
    "langchain-core>=0.3.0",

    # LangGraph
    "langgraph>=0.2.0",

    # pgvector / PostgreSQL memory
    "langchain-postgres>=0.0.12",
    "asyncpg>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.13.0",
    "pgvector>=0.3.0",
    "psycopg[binary]>=3.2.0",

    # Neo4j (RKG)
    "neo4j>=5.26.0",

    # Auth & crypto
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "cryptography>=43.0.0",

    # Observability
    "structlog>=24.0.0",

    # Utils
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "httpx>=0.27",
    "ruff>=0.8",
    "mypy>=1.11",
    "types-python-jose>=3.3.4",
    "types-passlib>=1.7.7",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
